+++ test/boost/sstable_test.cc
@@ -42,13 +42,10 @@
 #include "test/lib/test_services.hh"
 #include "cell_locking.hh"
 #include "sstables/sstable_mutation_reader.hh"
-#include "sstables/kl/reader_impl.hh"
 
 #include <boost/range/combine.hpp>
 
 using namespace sstables;
-using row_consumer = sstables::kl::row_consumer;
-using data_consume_rows_context = sstables::kl::data_consume_rows_context;
 
 bytes as_bytes(const sstring& s) {
     return { reinterpret_cast<const int8_t*>(s.data()), s.size() };
@@ -335,388 +332,6 @@ SEASTAR_TEST_CASE(compressed_random_access_read) {
         return sstables::test(sstp).data_read(env.make_reader_permit(), 97, 6).then([sstp] (temporary_buffer<char> buf) {
             BOOST_REQUIRE(sstring(buf.get(), buf.size()) == "gustaf");
             return make_ready_future<>();
-        });
-    });
-}
-
-class test_row_consumer : public row_consumer {
-public:
-    const int64_t desired_timestamp;
-    int count_row_start = 0;
-    int count_cell = 0;
-    int count_deleted_cell = 0;
-    int count_range_tombstone = 0;
-    int count_row_end = 0;
-
-    test_row_consumer(reader_permit permit, int64_t t)
-        : row_consumer(std::move(permit), tracing::trace_state_ptr()
-        , default_priority_class()), desired_timestamp(t) {
-    }
-
-    virtual proceed consume_row_start(sstables::key_view key, sstables::deletion_time deltime) override {
-        BOOST_REQUIRE(key == sstables::key_view(as_bytes("vinna")));
-        BOOST_REQUIRE(deltime.local_deletion_time == std::numeric_limits<int32_t>::max());
-        BOOST_REQUIRE(deltime.marked_for_delete_at == std::numeric_limits<int64_t>::min());
-        count_row_start++;
-        return proceed::yes;
-    }
-
-    virtual proceed consume_cell(bytes_view col_name, fragmented_temporary_buffer::view value,
-            int64_t timestamp, int64_t ttl, int64_t expiration) override {
-        BOOST_REQUIRE(ttl == 0);
-        BOOST_REQUIRE(expiration == 0);
-        switch (count_cell) {
-        case 0:
-            // The silly "cql marker" column
-            BOOST_REQUIRE(col_name.size() == 3 && col_name[0] == 0 && col_name[1] == 0 && col_name[2] == 0);
-            BOOST_REQUIRE(value.size_bytes() == 0);
-            BOOST_REQUIRE(timestamp == desired_timestamp);
-            break;
-        case 1:
-            BOOST_REQUIRE(col_name.size() == 7 && col_name[0] == 0 &&
-                    col_name[1] == 4 && col_name[2] == 'c' &&
-                    col_name[3] == 'o' && col_name[4] == 'l' &&
-                    col_name[5] == '1' && col_name[6] == '\0');
-            BOOST_REQUIRE(linearized(value) == as_bytes("daughter"));
-            BOOST_REQUIRE(timestamp == desired_timestamp);
-            break;
-        case 2:
-            BOOST_REQUIRE(col_name.size() == 7 && col_name[0] == 0 &&
-                    col_name[1] == 4 && col_name[2] == 'c' &&
-                    col_name[3] == 'o' && col_name[4] == 'l' &&
-                    col_name[5] == '2' && col_name[6] == '\0');
-            BOOST_REQUIRE(linearized(value) == bytes({0, 0, 0, 3}));
-            BOOST_REQUIRE(timestamp == desired_timestamp);
-            break;
-        }
-        count_cell++;
-        return proceed::yes;
-    }
-
-    virtual proceed consume_counter_cell(bytes_view col_name, fragmented_temporary_buffer::view value, int64_t timestamp) override {
-        BOOST_FAIL("counter cell wasn't expected");
-        abort(); // BOOST_FAIL is not marked as [[noreturn]].
-    }
-
-    virtual proceed consume_deleted_cell(bytes_view col_name, sstables::deletion_time deltime) override {
-        count_deleted_cell++;
-        return proceed::yes;
-    }
-
-    virtual proceed consume_shadowable_row_tombstone(bytes_view col_name, sstables::deletion_time deltime) override {
-        BOOST_FAIL("shdowable row tombstone wasn't expected");
-        abort(); // BOOST_FAIL is not marked as [[noreturn]].
-    }
-
-    virtual proceed consume_range_tombstone(
-            bytes_view start_col, bytes_view end_col,
-            sstables::deletion_time deltime) override {
-        count_range_tombstone++;
-        return proceed::yes;
-    }
-
-    virtual proceed consume_row_end() override {
-        count_row_end++;
-        return proceed::yes;
-    }
-
-    virtual void reset(indexable_element) override { }
-
-    virtual position_in_partition_view position() override {
-        return position_in_partition_view(position_in_partition_view::partition_start_tag_t{});
-    }
-};
-
-SEASTAR_TEST_CASE(uncompressed_rows_read_one) {
-    return test_using_reusable_sst(uncompressed_schema(), uncompressed_dir(), 1, [this] (auto& env, auto sstp) {
-        return do_with(test_row_consumer(env.make_reader_permit(), 1418656871665302), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c, {0, 95}, 95);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_cell == 3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
-        });
-    });
-}
-
-SEASTAR_TEST_CASE(compressed_rows_read_one) {
-    auto s = make_schema_for_compressed_sstable();
-    return test_using_reusable_sst(std::move(s), "test/resource/sstables/compressed", 1, [this] (auto& env, auto sstp) {
-        return do_with(test_row_consumer(env.make_reader_permit(), 1418654707438005), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c, {0, 95}, 95);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_cell == 3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
-        });
-    });
-}
-
-// Tests for iterating over all rows.
-
-class count_row_consumer : public row_consumer {
-public:
-    int count_row_start = 0;
-    int count_cell = 0;
-    int count_deleted_cell = 0;
-    int count_row_end = 0;
-    int count_range_tombstone = 0;
-
-    explicit count_row_consumer(reader_permit permit)
-        : row_consumer(std::move(permit), tracing::trace_state_ptr(), default_priority_class()) {
-    }
-
-    virtual proceed consume_row_start(sstables::key_view key, sstables::deletion_time deltime) override {
-        count_row_start++;
-        return proceed::yes;
-    }
-    virtual proceed consume_cell(bytes_view col_name, fragmented_temporary_buffer::view value,
-            int64_t timestamp, int64_t ttl, int64_t expiration) override {
-        count_cell++;
-        return proceed::yes;
-    }
-    virtual proceed consume_counter_cell(bytes_view col_name, fragmented_temporary_buffer::view value, int64_t timestamp) override {
-        count_cell++;
-        return proceed::yes;
-    }
-    virtual proceed consume_deleted_cell(bytes_view col_name, sstables::deletion_time deltime) override {
-        count_deleted_cell++;
-        return proceed::yes;
-    }
-    virtual proceed consume_row_end() override {
-        count_row_end++;
-        return proceed::yes;
-    }
-    virtual proceed consume_shadowable_row_tombstone(bytes_view col_name, sstables::deletion_time deltime) override {
-        count_deleted_cell++;
-        return proceed::yes;
-    }
-    virtual proceed consume_range_tombstone(
-            bytes_view start_col, bytes_view end_col,
-            sstables::deletion_time deltime) override {
-        count_range_tombstone++;
-        return proceed::yes;
-    }
-    virtual void reset(indexable_element) override { }
-    virtual position_in_partition_view position() override {
-        return position_in_partition_view(position_in_partition_view::partition_start_tag_t{});
-    }
-};
-
-
-SEASTAR_TEST_CASE(uncompressed_rows_read_all) {
-    return test_using_reusable_sst(uncompressed_schema(), uncompressed_dir(), 1, [this] (auto& env, auto sstp) {
-        return do_with(count_row_consumer(env.make_reader_permit()), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 4);
-                BOOST_REQUIRE(c.count_row_end == 4);
-                BOOST_REQUIRE(c.count_cell == 4*3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
-        });
-    });
-}
-
-SEASTAR_TEST_CASE(compressed_rows_read_all) {
-    auto s = make_schema_for_compressed_sstable();
-    return test_using_reusable_sst(std::move(s), "test/resource/sstables/compressed", 1, [this] (auto& env, auto sstp) {
-        return do_with(count_row_consumer(env.make_reader_permit()), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 4);
-                BOOST_REQUIRE(c.count_row_end == 4);
-                BOOST_REQUIRE(c.count_cell == 4*3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
-        });
-    });
-}
-
-// test reading all the rows one by one, using the feature of the
-// consume_row_end returning proceed::no message
-class pausable_count_row_consumer : public count_row_consumer {
-public:
-    using count_row_consumer::count_row_consumer;
-    virtual proceed consume_row_end() override {
-        count_row_consumer::consume_row_end();
-        return proceed::no;
-    }
-};
-
-SEASTAR_TEST_CASE(pausable_uncompressed_rows_read_all) {
-    return test_using_reusable_sst(uncompressed_schema(), uncompressed_dir(), 1, [this] (auto& env, auto sstp) {
-        return do_with(pausable_count_row_consumer(env.make_reader_permit()), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] () mutable {
-                // After one read, we only get one row
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_cell == 1*3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                auto fut = context->consume_input();
-                return fut.then([&c, context = std::move(context)] () mutable {
-                    // After two reads
-                    BOOST_REQUIRE(c.count_row_start == 2);
-                    BOOST_REQUIRE(c.count_row_end == 2);
-                    BOOST_REQUIRE(c.count_cell == 2*3);
-                    BOOST_REQUIRE(c.count_deleted_cell == 0);
-                    BOOST_REQUIRE(c.count_range_tombstone == 0);
-                    return make_ready_future<>();
-                    // FIXME: read until the last row.
-                });
-            });
-        });
-    });
-}
-
-
-// Test reading range tombstone (which we we have in collections such as set)
-class set_consumer : public count_row_consumer {
-public:
-    using count_row_consumer::count_row_consumer;
-    virtual proceed consume_range_tombstone(
-            bytes_view start_col, bytes_view end_col,
-            sstables::deletion_time deltime) override {
-        count_row_consumer::consume_range_tombstone(start_col, end_col, deltime);
-        // Note the unique end-of-component markers -1 and 1, specifying a
-        // range between start and end of row.
-        BOOST_REQUIRE(start_col == bytes({0, 9, 'f', 'a', 'v', 'o', 'r', 'i', 't', 'e', 's', -1}));
-        BOOST_REQUIRE(end_col == as_bytes({0, 9, 'f', 'a', 'v', 'o', 'r', 'i', 't', 'e', 's', 1}));
-        // Note the range tombstone have an interesting, not default, deltime.
-        BOOST_REQUIRE(deltime.local_deletion_time == 1428855312U);
-        BOOST_REQUIRE(deltime.marked_for_delete_at == 1428855312063524UL);
-        return proceed::yes;
-    }
-};
-
-SEASTAR_TEST_CASE(read_set) {
-    return test_using_reusable_sst(set_schema(), "test/resource/sstables/set", 1, [this] (auto& env, auto sstp) {
-        return do_with(set_consumer(env.make_reader_permit()), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_cell == 3);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_range_tombstone == 1);
-               return make_ready_future<>();
-            });
-        });
-    });
-}
-
-class ttl_row_consumer : public count_row_consumer {
-public:
-    const int64_t desired_timestamp;
-    ttl_row_consumer(reader_permit permit, int64_t t) : count_row_consumer(std::move(permit)), desired_timestamp(t) { }
-    virtual proceed consume_row_start(sstables::key_view key, sstables::deletion_time deltime) override {
-        count_row_consumer::consume_row_start(key, deltime);
-        BOOST_REQUIRE(key == sstables::key_view(as_bytes("nadav")));
-        BOOST_REQUIRE(deltime.local_deletion_time == std::numeric_limits<int32_t>::max());
-        BOOST_REQUIRE(deltime.marked_for_delete_at == std::numeric_limits<int64_t>::min());
-        return proceed::yes;
-    }
-
-    virtual proceed consume_cell(bytes_view col_name, fragmented_temporary_buffer::view value,
-            int64_t timestamp, int64_t ttl, int64_t expiration) override {
-        switch (count_cell) {
-        case 0:
-            // The silly "cql row marker" cell
-            BOOST_REQUIRE(col_name.size() == 3 && col_name[0] == 0 && col_name[1] == 0 && col_name[2] == 0);
-            BOOST_REQUIRE(value.size_bytes() == 0);
-            BOOST_REQUIRE(timestamp == desired_timestamp);
-            BOOST_REQUIRE(ttl == 3600);
-            BOOST_REQUIRE(expiration == 1430154618);
-            break;
-        case 1:
-            BOOST_REQUIRE(col_name.size() == 6 && col_name[0] == 0 &&
-                    col_name[1] == 3 && col_name[2] == 'a' &&
-                    col_name[3] == 'g' && col_name[4] == 'e' &&
-                    col_name[5] == '\0');
-            BOOST_REQUIRE(linearized(value) == bytes({0, 0, 0, 40}));
-            BOOST_REQUIRE(timestamp == desired_timestamp);
-            BOOST_REQUIRE(ttl == 3600);
-            BOOST_REQUIRE(expiration == 1430154618);
-            break;
-        }
-        count_row_consumer::consume_cell(col_name, value, timestamp, ttl, expiration);
-        return proceed::yes;
-    }
-};
-
-SEASTAR_TEST_CASE(ttl_read) {
-    return test_using_reusable_sst(uncompressed_schema(), "test/resource/sstables/ttl", 1, [this] (auto& env, auto sstp) {
-        return do_with(ttl_row_consumer(env.make_reader_permit(), 1430151018675502), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_cell == 2);
-                BOOST_REQUIRE(c.count_deleted_cell == 0);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
-        });
-    });
-}
-
-class deleted_cell_row_consumer : public count_row_consumer {
-public:
-    using count_row_consumer::count_row_consumer;
-    virtual proceed consume_row_start(sstables::key_view key, sstables::deletion_time deltime) override {
-        count_row_consumer::consume_row_start(key, deltime);
-        BOOST_REQUIRE(key == key_view(as_bytes("nadav")));
-        BOOST_REQUIRE(deltime.local_deletion_time == std::numeric_limits<int32_t>::max());
-        BOOST_REQUIRE(deltime.marked_for_delete_at == std::numeric_limits<int64_t>::min());
-        return proceed::yes;
-    }
-
-    virtual proceed consume_deleted_cell(bytes_view col_name, sstables::deletion_time deltime) override {
-        count_row_consumer::consume_deleted_cell(col_name, deltime);
-        BOOST_REQUIRE(col_name.size() == 6 && col_name[0] == 0 &&
-                col_name[1] == 3 && col_name[2] == 'a' &&
-                col_name[3] == 'g' && col_name[4] == 'e' &&
-                col_name[5] == '\0');
-        BOOST_REQUIRE(deltime.local_deletion_time == 1430200516);
-        BOOST_REQUIRE(deltime.marked_for_delete_at == 1430200516937621UL);
-        return proceed::yes;
-    }
-};
-
-SEASTAR_TEST_CASE(deleted_cell_read) {
-    return test_using_reusable_sst(uncompressed_schema(), "test/resource/sstables/deleted_cell", 2, [this] (auto& env, auto sstp) {
-        return do_with(deleted_cell_row_consumer(env.make_reader_permit()), [sstp] (auto& c) {
-            auto context = data_consume_rows<data_consume_rows_context>(*uncompressed_schema(), sstp, c);
-            auto fut = context->consume_input();
-            return fut.then([sstp, &c, context = std::move(context)] {
-                BOOST_REQUIRE(c.count_row_start == 1);
-                BOOST_REQUIRE(c.count_cell == 0);
-                BOOST_REQUIRE(c.count_deleted_cell == 1);
-                BOOST_REQUIRE(c.count_row_end == 1);
-                BOOST_REQUIRE(c.count_range_tombstone == 0);
-                return make_ready_future<>();
-            });
         });
     });
 }
